// src/constants/enums.ts
export enum Icons {
  PLUS = "plus",
  PLUS_CIRCLED_FILL = "plusCircleFill",
  COPY = "copy",
  PIN = "pin",
  LEFT_ARROW = "larrow",
  REACTION = "reaction",
  MICROPHONE = "mp",
  EARPHONES = "ep",
  GEAR = "gear",
  CROSS = "cross",
  CHAT = "chat",
}

export enum ButtonVariant {
  PRIMARY = "primary",
  SECONDARY = "secondary",
  GHOST = "ghost",
}